app/requirements.txt
Flask
pandas
numpy
psycopg2-binary
